requirements.txt
asgiref==3.5.2
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==4.0.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
executing==1.2.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==1.7.1
pytest==7.2.0
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dotenv==1.0.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.7.0
tzdata==2023.3
wcwidth==0.2.5
